EN-403: Correct ternary for creating a logging bucket (fix)
This resource should only be created when `var.logging_bucket_name` is
not specified.  If the user provides a name for the logging bucket then
we assume the bucket exists and use it.  Inverted logic to query for a
data source could be added to ensure that the bucket exists in a subsequent
enhancement.