EN-403: Add object-lock for WORM capabilities in s3 buckets (feat)
Enables the ability to set either GOVERNANCE or COMPLIANCE mode
Enables the ability to set the number of days of at least 1 day
for new objects created within the S3 bucket.

The rule can be made dynamic to allow for new objects to no longer
observe the default object-lock status.  We can revisit that if
there is any demand for it.  Allowing null for days can drive that
dynamic content block.