EN-403: Better support specifying logging_bucket_name
There were cases in which we would output that logging was not enabled
in the case that a logging_bucket_name was specified.  This will perform
a data lookup on the specified named and coalesce the resource with the data
in order to return the logging bucket arn/id which is then used in the
output for bucket_logging_enabled. The bucket_logging_arn output will only
exist if logging_bucket_arn has a non-empty value, otherwise the output
is set to null and terraform will not return it.