EN-878: Fix bug in `aws-cloudtrail-monitoring` module (fix)
Cloudwatch used to have a bug where it's log group ARN
contained a `:*` at the end. This was removed the 3.0.0
provider. However, there were still some other AWS
services that expect it; Cloudtrail is one of them.

We missed out one last ARN when we first created the
Cloudtrail module. This fixes that last ARN.